pyproject.toml
[project]
name = "2025-planning-quera"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "kirin-toolchain>=0.9.2",
    "bloqade[qasm2,pyqrack-cpu]>=0.16.0",
    "bloqade-hackathon-move>=0.1.0",
    "matplotlib>=3.10.0",
    "numpy>=2.2.2",
    "mqt-qcec>=2.8.1",
    "qiskit>=1.3.2",
    "iquhack-scoring>=0.1.0",
    "ipykernel>=6.29.5",
    "pyqt6>=6.8.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3.4",
]

[tool.uv.sources]
iquhack-scoring = { path = "./assets/scorer", editable = true }
